perf: store both QA models in half precision, 664 MB off the image
The generator shipped as a pickled fp32 `pytorch_model.bin` of 892 MB, the answer model as 437 MB of
safetensors. Measured in the image with the same inputs and fp32 arithmetic on both sides, rounding the
stored weights to fp16 changed nothing: 12 of 12 generated questions and 8 of 8 extracted answer spans came
back word for word. The image drops from 3.40 to 2.74 GB, and the pickle goes with it - safetensors needs no
`torch.load`.

Halved is the file, not the arithmetic. transformers 5 loads in the dtype of the file, so `load_qa_models`
now names `float32` explicitly; without it the service would have started computing in fp16, which measured
the same on this CPU but is not something this image can promise for every host.

Two numbers only the counter-check put right: the first load looked expensive at 10.5 s against 2.0 s, but
that was a cold cache on a freshly written file - warm, the conversion costs 0.8 s and there are half as many
bytes to read.

Verified in the image, not in the build log: no `.bin` left, 446 MB and 219 MB of safetensors, both models
load as float32, and all four smoke checks pass.

+++ app/synthesis/qa_models.py
@@ -152,10 +152,13 @@ def load_qa_models(qg_path: str, qa_path: str) -> QaModels | None:
         import torch  # optional extra "qa-models"; the service runs without it
         from transformers import AutoModelForQuestionAnswering, AutoModelForSeq2SeqLM, AutoTokenizer
 
+        # The weights are stored in half precision to halve what they weigh (scripts/fetch_qa_models.py),
+        # and transformers follows the dtype of the file. Naming float32 keeps the arithmetic where it was
+        # measured: fp16 on the CPU is not something this image can promise for every host it runs on.
         qg_tokenizer = AutoTokenizer.from_pretrained(qg_path)
-        qg_model = AutoModelForSeq2SeqLM.from_pretrained(qg_path)
+        qg_model = AutoModelForSeq2SeqLM.from_pretrained(qg_path, dtype=torch.float32)
         qa_tokenizer = AutoTokenizer.from_pretrained(qa_path)
-        qa_model = AutoModelForQuestionAnswering.from_pretrained(qa_path)
+        qa_model = AutoModelForQuestionAnswering.from_pretrained(qa_path, dtype=torch.float32)
     except Exception as exc:  # a missing package or a broken directory must not take the endpoint down
         log.error("QA models not usable (%s, %s): %s", qg_path, qa_path, exc)
         return None
